pyproject.toml
[project]
name = "pyapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "psycopg2-binary>=2.9.10",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "mypy>=1.17.0",
    "ruff>=0.12.5",
    "types-psycopg2>=2.9.21.20250718",
    "types-requests>=2.32.4.20250611",
]
parser = [
    "beautifulsoup4>=4.13.4",
    "requests>=2.32.4",
]
